+++ changelog.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## (unreleased)
+## v4.0.0 (unreleased)
 
 - structurizr-dsl: Fixes https://github.com/structurizr/java/issues/374 (!identifiers hierarchical isn't propagated when extending a workspace).
+- structurizr-dsl: Adds the ability to use the `group` keyword inside a component definition, to set the group name of that component.
 
 ## 3.2.1 (10th December 2024)